requirements.txt
opencv-python
requests
python-dotenv
pyautogui
pillow